feat: search product item control (#2543)
## What's the purpose of this pull request?

This PR is about the creation of a new component for SearchProduct:
SearchProductItemControl. Its goal is to display the QuantitySelector
and IconButton, while also fitting into the existing structure of
SearchProductItem.

The SearchProductItemControl will be part of the implementation of the
new "Quick Order" feature, which aims to allow users to place orders
quickly and efficiently.

## How it works?

The SearchProductItemControl will be used within the
SearchProductItemContent. A validation will check whether the QuickOrder
feature is enabled or not. If enabled, the SearchProductItemControl will
be rendered, maintaining the previous structure with the addition of the
controls: `QuantitySelector`, `CartButton`, and/or `SKUMatrix`.

The SearchProductItemControl will receive several props. Among them is
`skuMatrixControl`, which will be responsible for rendering the
`SKUMatrix`.

Preview: https://starter-b8rhu2zcu-vtex.vercel.app/

## Printscreens


![image](https://github.com/user-attachments/assets/c0d5cda3-ceec-4947-b794-d5aa8cd502d6)


![image](https://github.com/user-attachments/assets/4a9c3aa5-446f-4199-94bc-b1bc2fcacdf7)

ps: the "Select multiple" button is just a regular button used for
visualization purposes. Once the SKUMatrix is passed, the SKUMatrix
trigger will be displayed.

---------

Co-authored-by: Hiago Moreira <hiago.moreira@cubos.io>
Co-authored-by: Fanny Chien <fannychienn@gmail.com>