feat: [Delivery Promise] Default postal code (#2708)
## What's the purpose of this pull request?

Add in `discovery.config.js` the config that will be used for the new
feature Delivery Promise and update code related to validate session to
change the `postalCode` to the default one if it's null or empty.

## How it works?

If the session's `postalCode` is null or empty, it changes to the
initial/default one configured in `discovery.config.js` in the `session`
object.

## How to test it?

In the starter preview's `discovery.config.js` file, I've added `90210`
as the default postal code and enabled the deliveryPromise feature. Then
if you test the preview you should see this postal code in the IndexedDB
of the browser's devtools.
![Screenshot 2025-02-28 at 14 00
42](https://github.com/user-attachments/assets/26db75a2-339d-4eef-bf83-0c3cb6fab915)

You can also test that it'll be updated to the default postal code if
the postal code in the IDB is null or empty previously for some reason
(like if you had already accessed the store before this change). It can
be done by [editing the IDB through a script in the browser's
devtools](https://developer.chrome.com/docs/devtools/storage/indexeddb).
Here is the snippet I've used:
![Screenshot 2025-02-28 at 14 17
40](https://github.com/user-attachments/assets/b15de9e7-0b77-44c9-b999-cdaad5062663)
```
var connection = indexedDB.open('keyval-store', 1);

connection.onsuccess = (e) => {
    var database = e.target.result;
    var objectStore = database.transaction(['keyval'],'readwrite').objectStore('keyval');
    var getRequest = objectStore.get('fs::session');
    getRequest.onsuccess = (e) => {
        var data = e.target.result;
        data.postalCode = null;
        var updateRequest = objectStore.put(data, 'fs::session');
        updateRequest.onsuccess = (event) => {
            console.log(event.target.result);
        }
    }
}
```

https://github.com/user-attachments/assets/104164bf-80e8-4bec-bcd5-982f4fe60e04


### Starters Deploy Preview

https://storeframework-cm652ufll028lmgv665a6xv0g-8kzj40dcm.b.vtex.app/
([PR](https://github.com/vtex-sites/starter.store/pull/705))

## References

- [Jira task 1](https://vtex-dev.atlassian.net/browse/SFS-2199)
- [Jira task 2](https://vtex-dev.atlassian.net/browse/SFS-2200)
- [Delivery Promise RFC - related section
1](https://docs.google.com/document/d/15IUc7_sQ_MxM67TJJI4fqbnTyc9PVdy_jJb5M0iMhyE/edit?tab=t.0#heading=h.igtfl04jtxbb)
- [Delivery Promise RFC - related section
2](https://docs.google.com/document/d/15IUc7_sQ_MxM67TJJI4fqbnTyc9PVdy_jJb5M0iMhyE/edit?tab=t.0#heading=h.h45g42imthwf)

## Checklist

- [ ] For documentation changes, ping `@Mariana-Caetano` to review and
update (Or submit a doc request)