feat: search product control implementations (#2547)
## What's the purpose of this pull request?

This PR is about the implementation of the `Quick Order` feature. Two
other PRs were created, one related to the new SearchProductItemControl
component and another handling all the typing.

In this PR, the new component, typing, and some logic related to the
overall implementation of the feature are brought together. A small
adjustment was also made to the useBuyButton hook to prevent the
`SliderOver` from opening when an item is added to the cart.

Attention This PR has dependencies on PRs: #2543 and #2546.

ps: Once PR #2543 is merged, the number of changes in this PR will
decrease.

Link preview: https://starter-b8rhu2zcu-vtex.vercel.app/

## Printscreens


![image](https://github.com/user-attachments/assets/c0d5cda3-ceec-4947-b794-d5aa8cd502d6)


![image](https://github.com/user-attachments/assets/4a9c3aa5-446f-4199-94bc-b1bc2fcacdf7)

ps: the "Select multiple" button is just a regular button used for
visualization purposes. Once the SKUMatrix is passed, the SKUMatrix
trigger will be displayed.

---------

Co-authored-by: Hiago Moreira <hiago.moreira@cubos.io>
Co-authored-by: Fanny Chien <fannychienn@gmail.com>