feat: Add SEO specifics for PDP (#2730)
## What's the purpose of this pull request?

This PR adds SEO specifics (`title` and `description` templates) for
PDPs.

## How it works?

The store will be able to define both `title` and `description`
templates through the `discovery.config.js` file.

## How to test it?

Check if the metatags for` title` and `description` are being built
correctly on a PDP, that is with the shape defined at
`discovery.config.default.js` file.

### Starters Deploy Preview

vtex-sites/starter.store#724