feat: update next-seo to ^6.6.0 and adjust SEO settings (#2735)
This pull request includes several changes to the SEO configuration and
package dependencies. The most important changes are the addition of new
SEO-related properties in the configuration files and the update of the
`next-seo` package version.

### SEO Configuration Enhancements:
*
[`packages/core/discovery.config.default.js`](diffhunk://#diff-4da6f8e98efa58c9a40ce9b654b9bfeecb6813091999cdea3d2408679d1bf3d3R120-R121):
Added `noindex` and `nofollow` properties to the configuration.
*
[`packages/core/next-seo.config.ts`](diffhunk://#diff-44cee9f10a52af4534def49e35d22467e888a2689e6f535f4009801be8917cd7R8-R9):
Added `dangerouslySetAllPagesToNoFollow` and
`dangerouslySetAllPagesToNoIndex` properties to the SEO configuration.

### Package Dependency Update:
*
[`packages/core/package.json`](diffhunk://#diff-0b810c38f3c138a3d5e44854edefd5eb966617ca84e62f06511f60acc40546c7L74-R74):
Updated the `next-seo` package version from `^6.4.0` to `^6.6.0`.