feat: Manually disable `hideUnavailableItems` flag for PDP queries (#2747)
## What's the purpose of this pull request?

This PR intends to manually disable the `hideUnavailableItems` flag for
cases that the product is unavailable for certain regions.

## How it works?

it will manually pass `hideUnavailableItems: false` to the
`product_search` request to the IS. By doing that and only if the
merchant checks the option to show unavailable items on Catalog config
(Admin), the PDP will always be accessible, even though the product is
not available. So the store will need to handle the unavailable
information (unavailable labels, no prices and disable buy button).

## How to test it?

Currently we do not have a testing store with regionalization features,
but you can validate on Starter store if the pages are working as they
should with and without postal code being set.

## Starter Deploy Preview

vtex-sites/starter.store#738

## References

Check the **RFC FastStore PDP regionalization** for context.