feat: loads first page product gallery from Build Time (getStaticProps) + improve number of requests (+) (#2721)
## What's the purpose of this pull request?

This PR aims to fetch the first page from the PLPs using the server
side. This can improve SEO and performance.

It uses the same variables that are used in the first fetch from the
client side. If the values have changed 5 minutes from the build or from
the latest fetch of the first page, we refetch the first page and update
the values.

It also improves: 
- The cache mechanism from PLPs (deepEquals instead of stringify (and
===)).
- The URL params updates (fuzzy and operator) - there is no longer a
need to refetch to add these params.
- Decrease the number of requests made on the Client Side in the first
render (6 to 4).
- The cache and refresh logic for the first page.


## How to test it?

You can run locally:
I recommend to set the cache to 1 minute to test the requests from PLP
here:


https://github.com/vtex/faststore/blob/47fd98ef2e3380bc2b84f5ff7922068d75077d45/packages/core/src/sdk/product/useShouldFetchFirstPage.ts#L3

you can also test it in the preview link:
https://starter-7w8gjvlea-vtex.vercel.app/office

Observe the requests from the PLP: 

- just 1 request to 

http://localhost:3000/api/graphql?operationName=ClientProductGalleryQuery
since we don't refetch to include fuzzy and operator params.

- there is no current page request to 
http://localhost:3000/api/graphql?operationName=ClientManyProductsQuery
since this was moved to the Server Side.

- After 5 minutes from build/last client side request, you can see the
current page request Client Side that will refresh the data.
http://localhost:3000/api/graphql?operationName=ClientManyProductsQuery

- Observe the URL params from the PLPs with fuzzy, operator and all
other facets/params.
- Test the navigation history in the story.
- Test the Search page as well, since changes from PLP components can
effect Search page as well.

### Starters Deploy Preview
PR
- https://github.com/vtex-sites/starter.store/pull/717

Preview:

https://starter-mrmhyx2h3-vtex.vercel.app/office

---------

Co-authored-by: Larícia Mota <laricia.mota@vtex.com.br>