fix: channel's regionId and seller (#2768)
The PR https://github.com/vtex/faststore/pull/2749 introduced a bug:
when the store doesn't have the Delivery Promise feature enabled, the
channel `regionId` wasn't being updated.

The Checkout updates the session's `regionId` when we update the session
to add info about `geoCoordinates`, `country`, and `postalCode`. But we
were only doing that when the Delivery Promise feature was enabled.
This conditional was removed.

For the `sellers` info that we get from Checkout request, we only need
to make this request if there is a `seller` info inside the `channel`.
So a conditional was introduced to reduce the need of making this
request every time.

Check it with Delivery Promise enabled and disabled. On both flows the
session in the IndexedDB should be updated to add `regionId` inside the
channel when a `postalCode` is added in the RegionBar.

To check the `seller` info: an option is changing the config inside
`discovery.config`:
`channel:
'{"salesChannel":"1","regionId":"","hasOnlyDefaultSalesChannel":"true",
"seller":"storeframework01"}',`
and also set a `postalCode` there:
`postalCode: '90210',`

The session in the IndexedDB should contain the valid seller (example:
`storeframework01`) inside the channel if it was added (in config, for
example).

<!--- Add a link to a deploy preview from `starter.store` with this
branch being used. --->

<!--- Tip: You can get an installable version of this branch from the
CodeSandbox generated when this PR is created. --->

<!--- Spread the knowledge: is there any content you used to create this
PR that is worth sharing? --->

<!--- Extra tip: adding references to related issues or mentioning
people important to this PR may be good for the documentation and
reviewing process --->