feat: Add PDP redirects support (#2777)
## What's the purpose of this pull request?

This PR intends to add redirects support for PDPs in the same way we do
for PLPs.

See #2583 for further details.

## How it works?

It performs the redirect in the `getStaticProps` function when a certain
page is not found and there is a registered redirect for that. Before
rendering a `404 Not Found` result to the user, we will check if there
is a redirect associated to the pathname requested by the user.

## How to test it?

- Use the starter deploy preview;
- Navigate to `.../ergonomic/p` then you should be redirected to
`.../ergonomic-granite-mouse-57815628/p`;

### Starters Deploy Preview

vtex-sites/faststoreqa.store#764