feat: GA `view_item_list` event only when item is in viewport of `ProductGallery` (#2771)
## What's the purpose of this pull request?

**Suggestion: review this PR using the `Hide white space` option.**

This PR aims to send the `view_item_list` event only when the item is in
the viewport of `ProductGallery` instead of sending all items even when
not in the viewport.

## How it works?

It creates a new `ProductSentinel` component that uses the viewport
observer from the 'react-intersection-observer' lib and a debounce to
group all new products in the viewport.

We don't need to remove objects from the event or add the product that
was already in the viewport, only the new ones.

## How to test it?

You can run it locally by adding gtm_debug=true as a query string and
seeing the `dataLayer` object in the console while scrolling in the PLP.

Example:
http://localhost:3000/office?gtm_debug=true

Before: All PLP products were sent.
After: as the scroll was being performed and the items appeared in the
viewport, they were sent.

| Before | After |
|--------|--------|
|<img width="1189" alt="Screenshot 2025-04-08 at 20 53 44"
src="https://github.com/user-attachments/assets/b0c0b6b3-b17c-4815-8efb-0a0ebe476f64"
/>|<img width="1418" alt="Screenshot 2025-04-08 at 20 56 11"
src="https://github.com/user-attachments/assets/4fe1d976-45ca-4a7f-b24f-b0763f0eed4a"
/>|



### Starters Deploy Preview

PR
- https://github.com/vtex-sites/starter.store/pull/759

Preview
- https://starter-9o3z6fhnr-vtex.vercel.app/office?gtm_debug=true

---------

Co-authored-by: Lucas Feijó <lucas.portela@vtex.com>