feat: explicitly add allowRedirect to IS req (#2791)
## What's the purpose of this pull request?

Use the `allowRedirect` facet from IS. 

## How it works?

The default value would be `false` so it won't affect the IS products
return.
It'll be `true` for the request we make when checking if there is some
redirect configured for that search term on IS (it's configured through
Admin).
When there is a `redirect` value in the return of the IS req, the IS
doesn't return the list of products, so the `allowRedirect = true`
causes an issue for searches inside a carousel, for example.

## How to test it?

I've configured the landing page `/test-lari-lp-redirect` which has a
shelf. The shelf uses the `nike` search term. I've also configured the
redirect for the term `nike` to go to the `/technology` page.

- Without this PR the `/test-lari-lp-redirect` doesn't display the shelf
because the search is empty.
https://storeframework.vtex.app/test-lari-lp-redirect
<img width="1508" alt="Screenshot 2025-04-15 at 17 14 34"
src="https://github.com/user-attachments/assets/085cb322-e538-440c-831a-14820ac9f35e"
/>

- With the changes of this PR
- it uses `allowRedirect = false` so there are results for this search:
https://storeframework-cm652ufll028lmgv665a6xv0g-gd881p5dm.b.vtex.app/test-lari-lp-redirect
<img width="1503" alt="Screenshot 2025-04-15 at 17 08 55"
src="https://github.com/user-attachments/assets/87a6297f-4220-4dc5-a90b-3f20b472e8bc"
/>
- If you test searching for `nike` in the search bar it's redirected
normally to `/technology`.
https://storeframework-cm652ufll028lmgv665a6xv0g-gd881p5dm.b.vtex.app/s?q=nike


https://github.com/user-attachments/assets/98b4564a-74de-4a2a-af81-c0666cd5b5bc



### Starters Deploy Preview
https://github.com/vtex-sites/faststoreqa.store/pull/776

## References

- [Jira task](https://vtex-dev.atlassian.net/browse/SO-464)