fix: Allow control to `Checkbox`'s `checked` attribute (#2614)
## What's the purpose of this pull request?

Previously, the checkbox field could not be controlled due to the
destructuring of the props it receives. This issue was identified during
the development of the _Product Comparison_ feature. To fix this and
allow the checkbox to be controlled, the "checked" property was removed
from the destructuring so it can be managed.

---------

Co-authored-by: Lucas Feijó <lucasfjportela@gmail.com>