feat: order details query (#2800)
## What's the purpose of this pull request?

This PR aims to create the order details query (OMS) in
`@faststore/api`.
It also calls the query from the corresponding page in
`@faststore/core`, integrates the data, and removes the mocks.

## How it works?

It uses the new order details API to get the order and pass it to the
corresponding page.
This API uses the `VtexIdclientAutCookie` to check permissions, so you
must log in first.

## How to test it?

In localhost, you must replicate the `VtexIdclientAutCookie` created
after logging in.

1. Log in using an account (e.g., storeframework -
https://homebrewqa.fast.store/).
2. Place and order using the faststore > minicart > checkout.
3. Get the order ID created when you finish an order or use the admin >
orders.
4. You should see some order details on the page
`/account/orders/<your-order-id>`

### Starters Deploy Preview

TBD