fix: Search page blink (#2789)
## What's the purpose of this pull request?

This PR intends to fix the blinking content on Search page when
searching for terms.

## How it works?

We check if the query should be refetched and now we render the
`EmptySearch` component only for cases we don't have any product data to
show.

## How to test it?

- Do a search for terms like `ergonomic`, `apple` and `unbranded` and
check if the page content does not blink anymore.

### Starters Deploy Preview

vtex-sites/faststoreqa.store#775

---------

Co-authored-by: Lucas Feijó <lucas.portela@vtex.com>