feat: Check products availability when setting postal code (#2817)
## What's the purpose of this pull request?

This PR intends to handle the scenario when shoppers are setting postal
code and there is no available products for that specific location.

## How it works?

For all Delivery Promises scenarios, when shoppers are setting their
postal code an input error should be displayed if there is no available
products for that location.

## How to test it?

- If there is a testing postal code with no products available you can
use it on `RegionModal` or `RegionPopover`;
- Or for local environment you can change the condition at
`useSetLocation` hook to mock the `no products` scenario;

### Starters Deploy Preview

vtex-sites/faststoreqa.store#788

---------

Co-authored-by: Larícia Mota <laricia.mota@vtex.com.br>