feat: integration SKU matrix quick order (#2620)
## What's the purpose of this pull request?

This PR relates to the integration of SKU Matrix with Quick Order. New
configurations were made in the sections.json file, enabling the
configuration of the SKU Matrix, the definition of column names, and the
addition of new columns to the table. A minor adjustment was made to the
SKU Matrix abstraction component to allow the sending of product
information, making it possible to retrieve all variations of the
product and display them in the SKU Matrix.

Additionally, a logic was implemented to monitor the opening and closing
of the modal to inform the SearchInput component, due to the
useOnClickOutside method. Since the SKU Matrix is rendered within
SearchInputItem, clicking on the SKU Matrix modal was causing the
SearchDropdown and, consequently, the SKU Matrix to close. Therefore, a
logic was added to maintain the previous behavior of the SearchInput
while accommodating the SKU Matrix.

Note: This PR is dependent on #2546 and #2547. After the merge of #2546
and #2547 into the main branch of the functionality, the total changes
in this PR will decrease.

## Printscreens


![image](https://github.com/user-attachments/assets/6a262932-0852-4be9-b19e-9917b4ff0ab9)


![image](https://github.com/user-attachments/assets/33ea4d24-af0a-4629-801a-0d312a7f4922)


![image](https://github.com/user-attachments/assets/e23dad25-4181-4046-a3c0-c5c7cb0d3e55)


![image](https://github.com/user-attachments/assets/056bf032-443d-4ff1-b889-0f4d1600a28a)

---------

Co-authored-by: Hiago Moreira <hiago.moreira@cubos.io>