fix: remove the useOnClickOutside import breaking change (#2835)
## What's the purpose of this pull request?

When [this file
](https://github.com/vtex/faststore/pull/2716/files#diff-2ad0a54eceb2d4f46e9303b3120312373a87f561a8eed452a8886f034e83f7df)
was moved to the /components folder, it broke this type of import:
`import useOnClickOutside from "src/sdk/ui/useOnClickOutside";`, that
some clients use. We do not recommend this import, but since it's used,
we shouldn't break without a major release.

## How to test it?

It should build without problems if the store uses `import
useOnClickOutside from "src/sdk/ui/useOnClickOutside";`.
It can be tested in the core, in the
[SearchInput](https://github.com/vtex/faststore/blob/ed400bc738dd7d51e0a1fffce9c13ed843b90ee0/packages/core/src/components/search/SearchInput/SearchInput.tsx#L22)
instead of importing from `@faststore/ui` import from
`src/sdk/ui/useOnClickOutside`.