feat: list orders search/filters integration + temp front style (#2825)
## What's the purpose of this pull request?

This PR aims to implement the List Orders search/ filter integration. 
Part of this logic is complemented in the next PR
- https://github.com/vtex/faststore/pull/2829
- preview
https://sfj-b47aeb4--faststoreqa.preview.vtex.app/account/orders

If you want to test a more complete experience, use the PR mentioned
above.

## How it works?

It integrates with FastStore API and Commerce OMS APIs.

## How to test it?

- OMS API is not ready yet, but we are passing the corresponding filters
already.

Preview from the frontend PR
https://sfj-b47aeb4--faststoreqa.preview.vtex.app/account/orders