feat: selected tags and clear all (#2837)
## depends on
- https://github.com/vtex/faststore/pull/2829

## What's the purpose of this pull request?

This PR aims to implement the selected tags feature in the orders page
of my account list. It also implements the clear all button.

PS: as the status mapping is not completed yet, the labels from the tags
are not beautiful, but as we implement, the labels will become beautiful
(capitalized and with spaces instead of -).

## How it works?

When there are filters selected (status or dates), they appear in the
tags with options of remove and clear all.

![May-09-2025
16-07-53](https://github.com/user-attachments/assets/1a38d648-0ea8-409e-a0ee-657c0c9f1721)


## How to test it?

Select some filter in `FIlterSlider`, then remove the tags.

### Starters Deploy Preview
 TBD