feat: canceled statuses order status (#2846)
## What's the purpose of this pull request?

Handles `Canceled` and `Cancellation Requested` status on
MyAccountStatusCard

<!--- Considering the context, what is the problem we'll solve? Where in
VTEX's big picture our issue fits in? Write a tweet about the context
and the problem itself. --->

## How it works?

When the order has `Canceled` or `Cancellation Requested` status, it
must show the cancel status in the middle and "ghost" the others.

## How to test it?

<!--- Describe the steps with bullet points. Is there any external link
that can be used to better test it or an example? --->

### Starters Deploy Preview

- https://github.com/vtex-sites/faststoreqa.store/pull/805

## References


![image](https://github.com/user-attachments/assets/4be9bd5e-a79a-44ec-8c89-29793d1daf2d)

![image](https://github.com/user-attachments/assets/4147b7bb-b771-4168-ae7d-37c0188d8c9a)