feat: my account org drawer API (#2848)
## What's the purpose of this pull request?

This PR aims to implement the FastStore API calls to make Org Drawer
work as expected. It also anticipated that some commerce API calls that
will be used in the following tasks.

## How it works?

It makes the b2b fields available in the `validateSession` mutation.

## How to test it?

You can call the validateSession mutation with a B2B user. Reach out to
me so that I can help with this case.
It's better to test with the frontend PR:
- https://github.com/vtex/faststore/pull/2849