chore: enabling typescript declarationMap feature (#2862)
## What's the purpose of this pull request?

Enable a typescript feature that allows the IDE to follow library code
directly to the source definition instead of the declaration generated
by TS.

<!--- Considering the context, what is the problem we'll solve? Where in
VTEX's big picture our issue fits in? Write a tweet about the context
and the problem itself. --->

## How it works?

Typescript will generate together with the declaration files a sourceMap
indicating the sourceCode that was used to generate that declaration.
This sourcemap is used by tools to identify the real code instead the
declaration only.

<!--- Tell us the role of the new feature, or component, in its context.
Provide details about what you have implemented and screenshots if
applicable. --->

## How to test it?
Enters the branch and runs a `pnpm build --force` to skip turborepo
cache and generated a fresh build with the new property.
After that code imported between packages should go directly to the
sourceCode. Example: `import {} from '@faststore/components'` inside the
@faststore/core package when followed in vscode (comand + left click)
should redirect the dev to the typescript source file of that import
instead the declaration inside the dist folder.

<!--- Describe the steps with bullet points. Is there any external link
that can be used to better test it or an example? --->

### Starters Deploy Preview

<!--- Add a link to a deploy preview from `starter.store` with this
branch being used. --->

<!--- Tip: You can get an installable version of this branch from the
CodeSandbox generated when this PR is created. --->

## References

<!--- Spread the knowledge: is there any content you used to create this
PR that is worth sharing? --->

<!--- Extra tip: adding references to related issues or mentioning
people important to this PR may be good for the documentation and
reviewing process --->

## Checklist

<em>You may erase this after checking them all :wink:</em>

**PR Title and Commit Messages**

- [ ] PR title and commit messages follow the [Conventional
Commits](https://www.conventionalcommits.org/en/v1.0.0/) specification
- Available prefixes: `feat`, `fix`, `chore`, `docs`, `style`,
`refactor`, `ci` and `test`

**PR Description**

- [ ] Added a label according to the PR goal - `breaking change`, `bug`,
`contributing`, `performance`, `documentation`..

**Dependencies**

- [ ] Committed the `pnpm-lock.yaml` file when there were changes to the
packages

**Documentation**

- [ ] PR description
- [ ] For documentation changes, ping `@Mariana-Caetano` to review and
update (Or submit a doc request)



![image](https://github.com/user-attachments/assets/cb0130e4-71b9-40b3-ad84-af13ad64b66c)