feat: Shipping methods integration (#2847)
## What's the purpose of this pull request?

This PR contains the shipping methods integration with API.

We already have some of the shipping facets being rendered, but no
action was being made. After these changes, the filters should work
depending on the selected scenario.

## How it works?

The `usePickupPoints` hook is responsible to get the list of pickup
points available for the specific location in session. Based on that
list, we create the `pickup-in-point` facet on the client-side (it's not
delivered by the IS API) and inject on the facets list.

When the `pickup-in-point` is selected, a additional facet called
`pickupPoint` is also added on the facets list, then on the Search
client (`@faststore/api`) we transform it on query string param and send
it in the `product_search` IS request.

| Desktop | Mobile |
| - | - |
| ![Screenshot 2025-05-26 at 14 39
03](https://github.com/user-attachments/assets/782a45b5-d1bd-4967-9a7e-abe8c6f41ae7)
| ![Screenshot 2025-05-26 at 14 38
53](https://github.com/user-attachments/assets/078fbc83-2b89-411a-9533-415486f99e7b)
|

## How to test it?

- Use the following postal code: `50030-260`;
- Pay attention to the behavior of each different filter: each one must
send the correct parameters/variables in the query (check the `Network`
tab);
- Observe the behavior and visibility of the filters on both desktop and
mobile, the logic is similar but they have some different details that
can impact navigation;

### Starters Deploy Preview

vtex-sites/faststoreqa.store#812