feat: show empty state instead of table (#2913)
## What's the purpose of this pull request?

Put empty state inside the page and show it instead of the table. This
way we continue o show the filters ui

## How it works?

<!--- Tell us the role of the new feature, or component, in its context.
Provide details about what you have implemented and screenshots if
applicable. --->

## How to test it?

<!--- Describe the steps with bullet points. Is there any external link
that can be used to better test it or an example? --->

### Starters Deploy Preview

<!--- Add a link to a deploy preview from `starter.store` with this
branch being used. --->

<!--- Tip: You can get an installable version of this branch from the
CodeSandbox generated when this PR is created. --->

## References

<!--- Spread the knowledge: is there any content you used to create this
PR that is worth sharing? --->

<!--- Extra tip: adding references to related issues or mentioning
people important to this PR may be good for the documentation and
reviewing process --->


![image](https://github.com/user-attachments/assets/054696e8-03ac-4816-a19d-dfbbb68f6395)

![image](https://github.com/user-attachments/assets/cd20ad21-d2af-40fb-9399-02949d0d7f70)